+++ src/node-lib/curriculum-api-2023/queries/curriculumUnits/curriculumUnits.schema.ts
@@ -5,6 +5,8 @@ const curriculumUnitsSchema = z.object({
     .object({
       connection_prior_unit_description: z.string().nullable(),
       connection_future_unit_description: z.string().nullable(),
+      connection_future_unit_title: z.string().nullable(),
+      connection_prior_unit_title: z.string().nullable(),
       domain: z.string().nullable(),
       domain_id: z.number().nullable(),
       examboard: z.string().nullable(),
@@ -40,6 +42,8 @@ const curriculumUnitsSchema = z.object({
         z.object({
           connection_prior_unit_description: z.string().nullable(),
           connection_future_unit_description: z.string().nullable(),
+          connection_prior_unit_title: z.string().nullable(),
+          connection_future_unit_title: z.string().nullable(),
           title: z.string(),
           unitvariant_id: z.number(),
           lessons: z.array(